backend/requirements.txt
Django
djangorestframework
django-cors-headers
django-environ
django-crontab
requests
ebaysdk
pika